+++ Tools/MakePhotonPlots.C
@@ -31,7 +31,7 @@ int main(int argc, char* argv[])
     bool doPlots = true, doSave = true, doTuple = true, fromTuple = true, runOnCondor = false;
     string histFile = "", dataSets = "", sampleloc = AnaSamples::fileDir, plotDir = "plots";
     int nFiles = -1, startFile = 0, nEvts = -1;
-    double lumi = AnaSamples::luminosity;
+    float lumi = AnaSamples::luminosity;
     std::string sbEra = "SB_v1_2017";//"SB_v1_2017";
 
     while((opt = getopt_long(argc, argv, "pstfcH:D:N:M:E:P:L:S:", long_options, &option_index)) != -1)
@@ -114,11 +114,11 @@ int main(int argc, char* argv[])
     AnaSamples::SampleSet        ss("sampleSets.txt");
     AnaSamples::SampleCollection sc("sampleCollections.txt", ss);
 
-    const double zAcc = 1.0;
-    //    const double zAcc = 0.5954;
-    //    const double zAcc = 0.855;
-    const double znunu_mumu_ratio = 5.942;
-    const double znunu_ee_ratio   = 5.942;
+    const float zAcc = 1.0;
+    //    const float zAcc = 0.5954;
+    //    const float zAcc = 0.855;
+    const float znunu_mumu_ratio = 5.942;
+    const float znunu_ee_ratio   = 5.942;
 
     map<string, vector<AnaSamples::FileSummary>> fileMap;
