+++ Tools/ScaleFactors.h
@@ -4,14 +4,14 @@
 class ScaleFactors
 {
 public:
-    static double sf_norm0b()
+    static float sf_norm0b()
     {
       //return 1.079;
       //return 0.703341;      
       //return 0.9777071;
       return 0.945298;
     }
-    static double sfunc_norm0b()
+    static float sfunc_norm0b()
     {
       //return 0.082;
       //return 0.038727;
