+++ Tools/ScaleFactorsttBar.h
@@ -4,11 +4,11 @@
 class ScaleFactorsttBar
 {
 public:
-    static double sf_norm0b()
+    static float sf_norm0b()
     {
         return 1.000;
     }
-    static double sfunc_norm0b()
+    static float sfunc_norm0b()
     {
         return 0.142;
     }
